tests/wpt/tests/css/css-text-decor/text-underline-offset-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: parsing text-underline-offset with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-4/#underline-offset">
<meta name="assert" content="text-underline-offset supports the following values: auto | <length> | <percentage>">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-underline-offset", "auto");
test_valid_value("text-underline-offset", "-10px");
test_valid_value("text-underline-offset", "2001em");
test_valid_value("text-underline-offset", "-49em");
test_valid_value("text-underline-offset", "53px");
test_valid_value("text-underline-offset", "5%");
test_valid_value("text-underline-offset", "89%");
test_valid_value("text-underline-offset", "-30%");
test_valid_value("text-underline-offset", "187%");
test_valid_value("text-underline-offset", "calc(45% - 0.3em)");
test_valid_value("text-underline-offset", "calc(40em - 10px)");
test_valid_value("text-underline-offset", "calc(-13em + 50px)");
</script>
</body>
</html>
